fix Main Activity PreLoader, fix TextFields transacction Screen
