+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/screens/transactionScreen/TransactionScreen.kt
@@ -52,6 +52,7 @@ import ar.edu.unlam.mobile.scaffold.ui.components.category.CategoryDisplay
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
+@RequiresApi(Build.VERSION_CODES.O)
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun TransactionScreen(
@@ -158,19 +159,6 @@ fun TransactionScreen(
             }
         }
 
-//        Text(text = convertedValue)
-//
-//        Text(text = "Estoy en la pantalla ${TransactionType.values()[selectedTab]}")
-//
-//        Button(onClick = {
-//            viewModel.getCurrencyConversion(
-//                source = "${selectedCurrency.code}",
-//                target = "ARS",
-//                quantity = amount,
-//            )
-//        }) {
-//            Text(text = "Convertir")
-//        }
         Text(text = "Categorías")
         CategoryDisplay(
             categories = categories,
@@ -204,8 +192,9 @@ fun TransactionScreen(
             onClick = {
                 val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm")
                 val current = LocalDateTime.now().format(formatter)
-
                 viewModel.createNewTransaction(selectedTabState, selectedCategory, selectedCurrency, amount.toDouble(), current, description)
+                amount = ""
+                description = ""
             },
         ) {
             Text(text = "Agregar")
